move templates logic to a separate provider
